fix(database): resolve schema validation error on 'salario' column
- Corrigido o erro de inicialização do Spring Boot (SchemaManagementException).
- Sincronizado o tipo da coluna 'salario' no PostgreSQL (de float8 para NUMERIC(15,2)) para bater com o tipo BigDecimal da entidade Java.
- Aplicação inicializando corretamente no Docker (porta 8081).

+++ docker/init.sql
@@ -2,7 +2,7 @@ CREATE TABLE IF NOT EXISTS funcionarios (
                                             id UUID PRIMARY KEY,
                                             nome VARCHAR(100) NOT NULL,
     cargo VARCHAR(100) NOT NULL,
-    salario FLOAT NOT NULL,
+    salario NUMERIC(15,2) NOT NULL,
     data_admissao DATE NOT NULL,
     data_demissao DATE,
     ativo BOOLEAN NOT NULL DEFAULT TRUE,
